more interactive create_open_order_page
