--- lib/widgets/pages/deliver_page/deliver_remark_page.dart
@@ -1,5 +1,0 @@
-
-/// This is the page directed when the user (deliverer) has filled the 'Pick Up point', 'Order Closing
-/// Time' and 'ETA' on the prev page.
-/// This page requires the deliverer to do two things - 1. specify the number of dishes to deliver and
-/// 2. additional remarks about the delivery.
